BUG/MEDIUM: http-ana: Try to handle response before handling server abort
In the request analyser responsible to forward the request, we try to detect
the server abort to stop the request forwarding. However, we must be careful
to not block the response processing, if any. Indeed, it is possible to get
the response and the server abort in same time. In this case, we must try to
forward the response to the client first.

So to fix the issue, in the request analyser we no longer handle the server
abort if the response channel is not empty. In the end, the response
analyser is able to detect the server abort if it is relevant. Otherwise,
the stream will be woken up after the response forwarding and the server
abort should be handled at this stage.

This patch should be backported as far as 2.7 only because the risk of
breakage is high. And it is probably a good idea to wait a bit before
backporting it.

(cherry picked from commit d3e379b3ce8b0bf6bb167bb1f06268edb5406284)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>
(cherry picked from commit bea90bbf23cb6f594ee58517bd7b783124f93ade)
 [ad: adjuste context, remove non present if block]
Signed-off-by: Amaury Denoyelle <adenoyelle@haproxy.com>