MINOR: hlua: Set context's appctx when the lua socket is created
The lua socket's context referenced the owning appctx. It was set when the
appctx was initialized. It is now performed when the appctx is created. It
is a small change but this will be required to fix several issues with the
lua sockets.

(cherry picked from commit ed9333827a6bf3c20e010485da12724992d2de51)
Signed-off-by: Amaury Denoyelle <adenoyelle@haproxy.com>
(cherry picked from commit 75d1bfbd693cacc45a0ee2098be0ece43570ae5f)
Signed-off-by: Amaury Denoyelle <adenoyelle@haproxy.com>