BUG/MEDIUM: actions: always apply a longest match on prefix lookup
Many actions take arguments after a parenthesis. When this happens, they
have to be tagged in the parser with KWF_MATCH_PREFIX so that a sub-word
is sufficient (since by default the whole block including the parenthesis
is taken).

The problem with this is that the parser stops on the first match. This
was OK years ago when there were very few actions, but over time new ones
were added and many actions are the prefix of another one (e.g. "set-var"
is the prefix of "set-var-fmt"). And what happens in this case is that the
first word is picked. Most often that doesn't cause trouble because such
similar-looking actions involve the same custom parser so actually the
wrong selection of the first entry results in the correct parser to be
used anyway and the error to be silently hidden.

But it's getting worse when accidentally declaring prefixes in multiple
files, because in this case it will solely depend on the object file link
order: if the longest name appears first, it will be properly detected,
but if it appears last, its other prefix will be detected and might very
well not be related at all and use a distinct parser. And this is random
enough to make some actions succeed or fail depending on the build options
that affect the linkage order. Worse: what if a keyword is the prefix of
another one, with a different parser but a compatible syntax ? It could
seem to work by accident but not do the expected operations.

The correct solution is to always look for the longest matching name.
This way the correct keyword will always be matched and used and there
will be no risk to randomly pick the wrong anymore.

This fix must be backported to the relevant stable releases.

(cherry picked from commit 1e3422e6b086ff3317149c5ca0619732fe3e54bc)
Signed-off-by: Willy Tarreau <w@1wt.eu>
(cherry picked from commit 106008b6a6ce87d32f99267a2b73d67cf5e249b0)
Signed-off-by: Willy Tarreau <w@1wt.eu>
(cherry picked from commit 46fa871daec11b505d08fe46128e0588db70784a)
Signed-off-by: Willy Tarreau <w@1wt.eu>
(cherry picked from commit c82ef7829bf22177da50ae62b6fcf4b6c6f3dea8)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>