MEDIUM: quic: count quic_conn instance for maxconn
Increment actconn and check maxconn limit when a quic_conn is
instantiated. This is necessary because prior to this patch, quic_conn
instances where not counted. Global actconn was only incremented after
the handshake has been completed and the connection structure is
allocated.

The increment is done using increment_actconn() on INITIAL packet
parsing if a new connection is about to be created. If the limit is
reached, the allocation is cancelled and the INITIAL packet is dropped.

The decrement is done under quic_conn_release(). This means that
quic_cc_conn instances are not taken into account. This seems safe
enough because quic_cc_conn are only used for minimal usage.

The counterpart of this change is that maxconn must not be checked a
second time when listener_accept() is done over a QUIC connection. For
this, a new bind_conf flag BC_O_XPRT_MAXCONN is set for listeners when
maxconn is already counted by the lower layer. For the moment, it is
positionned only for QUIC listeners.

Without this patch, haproxy process could suffer from heavy memory/CPU
load if the number of concurrent handshake is high.

This patch is not considered a bug fix per-se. However, it has a major
benefit to protect against too many QUIC handshakes. As such, it should
be backported up to 2.6. For this, it relies on the following patch :
  "MINOR: frontend: implement a dedicated actconn increment function"

(cherry picked from commit 7735cf3854eb155a50a5ea747406f2a25657e25c)
[cf: quic_rx.c does not exist on 2.8. Patch applied to quic_conn.c instead]
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>
(cherry picked from commit 331b8b1cc87d2055e2ccdb6d85fc7e731002d94e)
 [ad: adjusted convert BC_O_UNLIMITED to LI_O_UNLIMITED usage inside
  listener_uses_maxconn(). Also remove increment of non-existing
  global maxconn_reached.]
Signed-off-by: Amaury Denoyelle <adenoyelle@haproxy.com>
(cherry picked from commit 67b1b759a89891e43ef0a5d1f22a1f55b1b19cd4)
[cf: ctx adjt]
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>