DEBUG: add a tainted flag when ha_panic() is called
This will make it easier to know that the panic function was called,
for the occasional case where the dump crashes and/or the stack is
corrupted and not much exploitable. Now at least it will be sufficient
to check the tainted value to know that someone called ha_panic(), and
it will also be usable to condition extra analysis.

(cherry picked from commit 46bbb3a33be4be1cf8a54b39dbd57433501229e8)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>