BUG/MINOR: mux-quic: fix early close if unset client timeout
When no client timeout is defined in the configuration, QCC timeout task
is never allocated. However, a NULL timeout task is also used as a
criteria in qcc_is_dead() to consider that the MUX instance should be
released as timeout stroke earlier.

This bug causes every connection to be closed by haproxy side with a
CONNECTION_CLOSE. This is notable when using several streams per
connection with only the first stream completed and the others failed.

To fix this, change timeout task allocation policy. It is now always
allocated. This means that if no timeout is defined, it will never be
run. This is not considered a waste of resource as no timeout in the
configuration is considered as an exception case. However, this has the
advantage to simplify the rest of the code which can now check for the
task instance without having an extra check on the timeout value.

This bug is labelled as minor as it only occurs if no timeout client is
defined which reports warning on startup as it may caused unexpected
behavior.

This bug should be backported up to 2.6.

(cherry picked from commit 47ed1181f29a56ccb04e5b80ef4f9414466ada7a)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>
(cherry picked from commit ca7a063a3ca4a6e2ba53b0238ca9a1d6953d6b0c)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>
(cherry picked from commit 65c8ccd11355912a1ad81b954d17d50e07c90dc6)
[cf: ctx adjt]
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>