BUG/MINOR: tcpcheck: Report hexstring instead of binary one on check failure
When an expect rule failed for a tcp-check, information about the expect
rule is dumped in the report. For a check on a binary string, a hexstring is
used in the configuration but the decoded string is dumped. It is an problem
because it can contain special characters. And it is not really handy
because there is no correspondance with the config.

So, now, the hexstring is dumped in the report. This way, we are sure there
is no special characters and it is easy to find it in the configuration.

This patch shoudl solve the issue #2326. It must be backported as far as
2.2.

(cherry picked from commit c72ab1cc6d8205152fb5accd929d4bd871871106)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>
(cherry picked from commit a8b082d92eedde8accb46cea1b853de52eb42547)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>
(cherry picked from commit cdaea09c36394c81d5724f1fb82ab40fd56619ca)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>